Create LICENSE
MIT License is compatible with our project because it is an open-source software license. We believe that other people may as well give their contribution to this app to make it more robust. This type of license permits unrestricted use, distribution, and reproduction in any medium, provided that attribution is preserved.